+++ r-sql.tf
@@ -13,12 +13,12 @@ resource "azurerm_mssql_server" "sql" {
   administrator_login          = var.administrator_login
   administrator_login_password = var.administrator_password
   dynamic "azuread_administrator" {
-    for_each = var.azuread_administrator != null ? ["azuread_administrator"] : []
+    for_each = var.azuread_administrator != null ? ["enabled"] : []
     content {
-      login_username              = lookup(var.azuread_administrator, "login_username")
-      object_id                   = lookup(var.azuread_administrator, "object_id")
-      tenant_id                   = lookup(var.azuread_administrator, "tenant_id")
-      azuread_authentication_only = lookup(var.azuread_administrator, "azuread_authentication_only")
+      login_username              = var.azuread_administrator.login_username
+      object_id                   = var.azuread_administrator.object_id
+      tenant_id                   = var.azuread_administrator.tenant_id
+      azuread_authentication_only = var.azuread_administrator.azuread_authentication_only
     }
   }
 
@@ -30,7 +30,7 @@ resource "azurerm_mssql_server" "sql" {
 }
 
 resource "azurerm_mssql_firewall_rule" "firewall_rule" {
-  count = length(var.allowed_cidr_list)
+  count = try(length(var.allowed_cidr_list), 0)
 
   name      = "rule-${count.index}"
   server_id = azurerm_mssql_server.sql.id
@@ -70,7 +70,7 @@ resource "azurerm_mssql_elasticpool" "elastic_pool" {
 }
 
 resource "azurerm_sql_virtual_network_rule" "vnet_rule" {
-  for_each            = { for subnet in local.allowed_subnets : subnet.name => subnet }
+  for_each            = try({ for subnet in local.allowed_subnets : subnet.name => subnet }, {})
   name                = each.key
   resource_group_name = var.resource_group_name
   server_name         = azurerm_mssql_server.sql.name