+++ variables-naming.tf
@@ -1,9 +1,23 @@
+# Generic naming variables
 variable "name_prefix" {
   description = "Optional prefix for the generated name"
   type        = string
   default     = ""
 }
 
+variable "name_suffix" {
+  description = "Optional suffix for the generated name"
+  type        = string
+  default     = ""
+}
+
+variable "use_caf_naming" {
+  description = "Use the Azure CAF naming provider to generate default resource name. `server_custom_name` and `elastic_pool_custom_name` override this if set. Legacy default name is used if this is set to `false`."
+  type        = bool
+  default     = true
+}
+
+# Custom naming override
 variable "server_custom_name" {
   description = "Name of the SQL Server, generated if not set."
   type        = string
@@ -14,4 +28,10 @@ variable "elastic_pool_custom_name" {
   description = "Name of the SQL Elastic Pool, generated if not set."
   type        = string
   default     = ""
+}
+
+variable "use_caf_naming_for_databases" {
+  description = "Use the Azure CAF naming provider to generate databases names."
+  type        = bool
+  default     = false
 }